rename and refactor the main nav related code
