+++ src/components/sidebar/ContentWithMainNav.sc.js
